pyproject.toml
[project]
name = "sitewatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "resend>=2.19.0",
]
